Moving stub to application
