+++ src/functions/sendCandidateResults/application/service/send-email.ts
@@ -1,6 +1,7 @@
 import { NotifyClient } from 'notifications-node-client';
 import { DocumentsServiceError } from '../../domain/errors/documents-service-error';
 import { EmailPersonalisation } from '../../domain/personalisation.model';
+import { NotifyClientStub } from '../stub/notify-client-stub';
 
 export async function sendEmail(
     emailAddress: string,
@@ -8,7 +9,7 @@ export async function sendEmail(
     personalisation: EmailPersonalisation,
     reference: string,
     emailReplyToId: string,
-    client: NotifyClient,
+    client: NotifyClient | NotifyClientStub,
 )  : Promise <any> {
 
   try {
